+++ apps/e-learning-app/src/app/components/ContentGrid.js
@@ -1,5 +1,5 @@
 import Video from './video/Video';
-import {Container, Row,Col} from 'react-bootstrap';
+import {Row,Col} from 'react-bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 const VideosRow = (props) => {
