+++ apps/todos.js
@@ -4,23 +4,23 @@
 
 // TODO: fix sidebar not extending to the bottom of the page V
 
-// TODO: add a folder for each component and separate the style from functionality X
+// TODO: add a folder for each component and separate the style from functionality V
 
 // TODO: separate pages from components V
 
 // TODO: Make components a bit more modular - videoContent, sidebar ? 
 
-// TODO: resolve warnings - sidebar, video X
+// TODO: resolve warnings - sidebar, video  V
 
 // TODO: add comments X
 
 // TODO: refactor for react bootstrap V
 
-// TODO: rename some files - VideoContent -> ModuleContent; VideosRow -> ContentGrid/List, X
-// SidebarData -> MockData ( and put everything in it ) X
+// TODO: rename some files - VideoContent -> ModuleContent; VideosRow -> ContentGrid/List, V
+// SidebarData -> MockData ( and put everything in it ) V
 
 // TODO: make the videosRow a responsive grid -> https://react-bootstrap.github.io/layout/grid/ V
 
-// TODO: separate mocked data from real data, like in sidebarData X
+// TODO: separate mocked data from real data, like in sidebarData  V
 
 // TODO: delete this file once all changes are applied X
